repo/requirements.txt
fastapi
uvicorn
psycopg2-binary
qdrant-client
nomic
groq
einops
sentence-transformers
sqlalchemy
requests

python-dotenv
httpx